refactor(validate): split 1010-line validate.rs into focused submodules
